backend/requirements.txt
openai>=1.12.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
typing-extensions>=4.9.0
